requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
google-generativeai
chromadb
